refactor: moved new table components to separate files
